requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-cloud-language==2.11.1
google-cloud-firestore==2.13.1
firebase-admin==6.2.0
python-jose==3.3.0
pydantic==2.4.2
python-multipart==0.0.6
requests==2.31.0
openai==1.3.0 